+++ packages/dialect-mysql/src/index.ts
@@ -6,5 +6,6 @@ export { MySqlJsonPathTranslator } from './json/JsonPathTranslator';
 export * from './MySqlDdlStrategy';
 export * from './MysqlDialect';
 export * from './MysqlOptionsBuilder';
+export * from './MySqlTypeMapper';
 export * from './sequenceEmulation';
 export * from './spatial-functions';