+++ packages/plugin-multi-tenant/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ts-linq/plugin-multi-tenant
 
+## 2.0.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@5.1.0
+  - @ts-linq/metadata@4.1.12
+
 ## 2.0.33
 
 ### Patch Changes