+++ packages/transformer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ts-linq/transformer
 
+## 2.1.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@5.1.0
+  - @ts-linq/ast@3.2.13
+
 ## 2.1.33
 
 ### Patch Changes